requirements.txt
pandas
fhir.resources